refactor: move core modules into utils package
