+++ pages/Library.py
@@ -5,8 +5,8 @@ import sys
 import streamlit as st
 st.set_page_config(page_title="ASK Auxiliary Source of Knowledge", initial_sidebar_state="collapsed")
 import pandas as pd
-from registry import load_registry_and_date
-import ui_utils
+from utils.registry import load_registry_and_date
+from utils import ui_utils
 
 
 sys.path.insert(0, ui_utils.parent_dir)
