+++ sidebar.py
@@ -1,6 +1,6 @@
 # sidebar.py
 import streamlit as st
-import ui_utils
+from utils import ui_utils
 
 __all__ = ["build_sidebar"]          # so autoflake/ruff know what’s public
